Java halt to stdout
