+++ Source/DafnyCore/Compilers/Compiler-js.cs
@@ -2498,13 +2498,7 @@ namespace Microsoft.Dafny.Compilers {
         }
         nodeProcess.StandardInput.Flush();
         nodeProcess.StandardInput.Close();
-        // Fixes a problem of Node on Windows, where Node does not prints to the parent console its standard outputs.
-        var errorProcessing = Task.Run(() => {
-          PassthroughBuffer(nodeProcess.StandardError, Console.Error);
-        });
-        PassthroughBuffer(nodeProcess.StandardOutput, Console.Out);
-        nodeProcess.WaitForExit();
-        errorProcessing.Wait();
+        PassOnOutput(nodeProcess);
         return nodeProcess.ExitCode == 0;
       } catch (System.ComponentModel.Win32Exception e) {
         outputWriter.WriteLine("Error: Unable to start node.js ({0}): {1}", psi.FileName, e.Message);
@@ -2512,13 +2506,5 @@ namespace Microsoft.Dafny.Compilers {
       }
     }
 
-    // We read character by character because we did not find a way to ensure
-    // final newlines are kept when reading line by line
-    void PassthroughBuffer(StreamReader input, TextWriter output) {
-      int current;
-      while ((current = input.Read()) != -1) {
-        output.Write((char)current);
-      }
-    }
   }
 }
